複数人プレイ時のエラーを修正

+++ Assets/Scripts/Character.cs
@@ -47,7 +47,7 @@ public class Character : MonoBehaviourPunCallbacks, IPunObservable
         else
         {
             TeamID = (byte)stream.ReceiveNext();
-            HP = (byte)stream.ReceiveNext();
+            HP = (short)stream.ReceiveNext();
         }
     }
